Work in progress (backup)
The code here should not be seen as complete or even as an indication of
future structure.

This might be an over-engineered soltution but it feels a bit nicer
(possibly with a MASSIVE bug in it too)

Instead of having two buffers and copying tokens into the buffer on each
iteration, to re-build the path one file at a time in order to run
"access" on it, I have settled on a more direct string manipulation
strategy.

Essentially, knowing how a string is processed by most string
manipulating libraries, I can get away with removing and replacing
'/' characters in order to manipulate how a printf or scanf would
interact with it.

Probably got a bit too excited over this